It's finally done. (For now at least)
This commit removes some of the features that won't see many uses. Improved the features from sbacprofile_v4, and implemented features that has to be added since sbacprofile_v4.

Most of the UI elements are now built off of Daisy UI instead of MUI. MUI is still included for the date pickers (I'm incapable of making one. :skull:).

Most of the errors that the user will encounter is happened because there's a problem with the object that's fetched from the API. I decided to handle the error by putting the status of the fetch within the state, and components will check a status before rendering data. This will prevent app-breaking errors (Hopefully). This has also given me an ability to add a skeleton loader for many components.

Changes:
- Home page has been removed.
- Announcements page, and its related components and functions has been removed.
- Request forms has also been removed.
- Some UI element behaviours has been changed. (modal, text fields)